requirements.txt
gpflow==1.5.0
Keras==2.1.5
loguru==0.5.3
numpy==1.19.2
pandas==1.1.5
py-cpuinfo==8.0.0
pystan==2.19.1.1
scikit-learn==0.24.1
tensorflow==1.12.0
mgwr==2.1.2
pyshp==2.1.3
openpyxl==3.0.7
overpy==0.6
geopy==2.2.0
